requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitarray==2.8.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
cytoolz==0.12.2
dataclasses-json==0.6.1
deep-translator==1.11.4
dnspython==2.4.2
elastic-transport==8.4.1
elasticsearch==8.10.0
emoji==2.8.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.1
filelock==3.12.4
filetype==1.2.0
Flask==3.0.0
frozenlist==1.4.0
fsspec==2023.9.2
fuzz==0.1.1
google-search-results==2.4.2
gptcache==0.1.42
greenlet==3.0.0
h11==0.14.0
hexbytes==0.3.1
hiredis==2.2.3
huggingface-hub==0.17.3
idna==3.4
iso-639==0.4.5
iso639==0.1.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
langchain==0.0.308
langchainhub==0.1.13
langdetect==1.0.9
langsmith==0.0.42
lingua-language-detector==1.3.3
lru-dict==1.2.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.24.4
openai==0.28.1
outcome==1.3.0.post0
packaging==23.2
parsimonious==0.9.0
pdfminer==20191125
pdfminer.six==20221105
polib==1.2.0
protobuf==4.25.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.11
pydantic_core==2.3.0
pymongo==4.5.0
PySocks==1.7.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-ulid==1.1.0
pyunormalize==15.1.0
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.3.1
redis==4.6.0
redis-om==0.2.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.13.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
tabulate==0.9.0
tenacity==8.2.3
the==0.1.5
thefuzz==0.20.0
tiktoken==0.5.1
toolz==0.12.0
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.7
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.10.30
urllib3==2.0.7
web3==6.11.3
websockets==12.0
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.9.2
